backend/requirements.txt
Flask==2.1.0
Flask-CORS==3.0.10
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.31
stripe==2.65.0
python-dotenv==0.19.2
gunicorn==20.1.0
pytest==7.1.1